Add files via upload
To solve this problem, we can iterate through the array of prices and look for opportunities to buy and sell the stock such that we maximize our profit. Since we are allowed to buy and sell on the same day and we can only hold at most one share of the stock at any time, we can simply add to our profit every time we see that the price of the stock has increased compared to the previous day.

Here's a step-by-step algorithm:

Initialize a variable maxProfit to 0. This will keep track of the maximum profit we can achieve.
Loop through the array starting from the second element because we can't sell on the first day.
For each day, check if the price of the stock has increased compared to the previous day.
If it has, add the difference to maxProfit.
Return maxProfit after the loop ends